avance

+++ CreativaSL.Web.Ganados/Models/System/UploadImageToServerModel.cs
@@ -11,6 +11,6 @@ namespace CreativaSL.Web.Ganados.Models.System
         public Exception Exception { get; set; }
         public bool Success { get; set; }
         public string UrlComplete { get; set; }
-        public int QualityImage { get; set; } = 50;
+        public long QualityImage { get; set; } = 50L;
     }
 }
